Added subtask entities and repos.
