+++ src/Entity/TaskBase.php
@@ -8,7 +8,7 @@ use Doctrine\ORM\Mapping as ORM;
 /**
  * @ORM\MappedSuperclass()
  */
-class TaskBase
+class TaskBase extends TaskFragmentBase
 {
     /**
      * @ORM\Column(type="date")
@@ -21,16 +21,6 @@ class TaskBase
      * @var Guardian
      */
     protected $guardian;
-    /**
-     * @ORM\Column(type="text")
-     * @var string
-     */
-    protected $imgUrl;
-    /**
-     * @ORM\Column(type="text")
-     * @var string
-     */
-    protected $text;
     /**
      * @ORM\Column(type="time")
      * @var DateTime
@@ -59,22 +49,6 @@ class TaskBase
     }
 
     /**
-     * @return string
-     */
-    public function getImgUrl(): string
-    {
-        return $this->imgUrl;
-    }
-
-    /**
-     * @return string
-     */
-    public function getText(): string
-    {
-        return $this->text;
-    }
-
-    /**
      * @return DateTime
      */
     public function getTimeEnd(): DateTime
@@ -92,7 +66,7 @@ class TaskBase
 
     /**
      * @param DateTime $day
-     * @return TaskBase
+     * @return $this
      */
     public function setDay(DateTime $day): TaskBase
     {
@@ -102,7 +76,7 @@ class TaskBase
 
     /**
      * @param Guardian $guardian
-     * @return TaskBase
+     * @return $this
      */
     public function setGuardian(Guardian $guardian): TaskBase
     {
@@ -111,28 +85,8 @@ class TaskBase
     }
 
     /**
-     * @param string $imgUrl
-     * @return TaskBase
-     */
-    public function setImgUrl(string $imgUrl): TaskBase
-    {
-        $this->imgUrl = $imgUrl;
-        return $this;
-    }
-
-    /**
-     * @param string $text
-     * @return TaskBase
-     */
-    public function setText(string $text): TaskBase
-    {
-        $this->text = $text;
-        return $this;
-    }
-
-    /**
      * @param DateTime $timeEnd
-     * @return TaskBase
+     * @return $this
      */
     public function setTimeEnd(DateTime $timeEnd): TaskBase
     {
@@ -142,7 +96,7 @@ class TaskBase
 
     /**
      * @param DateTime $timeStart
-     * @return TaskBase
+     * @return $this
      */
     public function setTimeStart(DateTime $timeStart): TaskBase
     {
