+++ src/Repository/TaskBaseRepository.php
@@ -2,8 +2,6 @@
 
 namespace App\Repository;
 
-use App\Component\Common\Service\BaseManager;
-
-abstract class TaskBaseRepository extends BaseRepository
+abstract class TaskBaseRepository extends TaskFragmentBaseRepository
 {
 }
