+++ src/Repository/TaskFragmentBaseRepository.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace App\Repository;
+
+abstract class TaskFragmentBaseRepository extends BaseRepository
+{
+}
